% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dna2sig}
\alias{dna2sig}
\title{Convert DNA sequence into numeric signal.}
\usage{
dna2sig(sequence)
}
\arguments{
\item{sequence}{DNAString or DNAStringSet object.}
}
\value{
Numeric vector with signal
}
\description{
Convert DNA sequence into numeric signal.
}
\examples{
sig <- dna2sig(DNAString("CCCCCCGGGTGGGTGGGTGGGAAAA"))

}

