% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{dna_signal}
\alias{dna_signal}
\title{Convert DNA sequence into numeric signal.}
\usage{
dna_signal(subject)
}
\arguments{
\item{subject}{DNAString or DNAStringSet object.}
}
\value{
Numeric vector with signal.
}
\description{
Convert DNA sequence into numeric signal.
}
\examples{
sig <- dna_signal(DNAString("CCCCCCGGGTGGGTGGGTGGGAAAA"))
sig_list <- dna_signal(DNAStringSet(c("ACGT","ACCT","GAAC")))

}

