% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigbin.R
\name{fasta_sigbin}
\alias{fasta_sigbin}
\title{Perform binning based on Hjorth parameters.}
\usage{
fasta_sigbin(filepath, k = NULL)
}
\arguments{
\item{filepath}{Path to FASTA file.}

\item{k}{Number of clusters.}
}
\value{
Input data frame extended by binning annotation.
}
\description{
Perform binning based on Hjorth parameters.
}
\examples{
file <- system.file("extdata", "reads.fasta", package="sigbin")
binning <- fasta_sigbin(file)

}

