% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigbin.R
\name{hjorth_params}
\alias{hjorth_params}
\title{Compute Hjorth parameters for DNA sequences or numeric signal.}
\usage{
hjorth_params(x)
}
\arguments{
\item{x}{DNAString, DNAStringSet, numeric vector or list of numeric vectors.}
}
\value{
Data frame of Hjorth parameters.
}
\description{
Compute Hjorth parameters for DNA sequences or numeric signal.
}
\examples{
sig_list <- dna_signal(DNAStringSet(c("ACGT","ACCT","GAAC")))
hp <- hjorth_params(sig_list)

}

