% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigbin.R
\name{sigbin}
\alias{sigbin}
\title{Perform binning based on Hjorth parameters.}
\usage{
sigbin(hp, k = NULL)
}
\arguments{
\item{hp}{Data frame of Hjorth parameters.}

\item{k}{Number of clusters.}
}
\value{
Input data frame extended by binning annotation.
}
\description{
Perform binning based on Hjorth parameters.
}
\examples{
file <- system.file("extdata", "reads.fasta", package="sigbin")
hp <- fasta_hjorth_params(file)
binning <- sigbin(hp)

}

