% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigbin.R
\name{plot_sigbin_3d}
\alias{plot_sigbin_3d}
\title{Plot signal based binning in 3D.}
\usage{
plot_sigbin_3d(sb)
}
\arguments{
\item{sb}{Data frame with binning annotation.}
}
\description{
Plot signal based binning in 3D.
}

